refactor(cv): tighten wallet pt.1 view tests and relocate OptinMetrics coverage
Move onboarding analytics CV tests to OptinMetrics, strengthen interaction assertions, and use named selectors for action-sheet confirmation.
