+++ app/components/Views/Settings/AppInformation/AppInformation.view.test.tsx
@@ -1,18 +1,53 @@
 import '../../../../../tests/component-view/mocks';
+import React from 'react';
 import { fireEvent, waitFor } from '@testing-library/react-native';
+import { InteractionManager } from 'react-native';
 import { describeForPlatforms } from '../../../../../tests/component-view/platform';
-import { renderComponentViewScreen } from '../../../../../tests/component-view/render';
+import {
+  getRouteProbeTestId,
+  renderScreenWithRoutes,
+} from '../../../../../tests/component-view/render';
 import { initialStateIdentity } from '../../../../../tests/component-view/presets/identity';
 import AppInformation from './index';
 import { AboutMetaMaskSelectorsIDs } from './AboutMetaMask.testIds';
 import { strings } from '../../../../../locales/i18n';
 import Routes from '../../../../constants/navigation/Routes';
 
+/**
+ * Component View tests for AppInformation (About MetaMask).
+ *
+ * Mirrors (partial): tests/smoke-appium/settings/contact-us.spec.ts
+ * — support links render and navigate to the in-app webview.
+ *
+ * Run: yarn jest -c jest.config.view.js AppInformation.view.test.tsx --runInBand
+ */
+
+beforeAll(() => {
+  jest.spyOn(InteractionManager, 'runAfterInteractions').mockImplementation(
+    jest.fn().mockImplementation((callback) => {
+      if (typeof callback === 'function') {
+        callback();
+      }
+      return {
+        then: (onfulfilled?: () => void) => Promise.resolve(onfulfilled?.()),
+        done: (onfulfilled?: () => void, onrejected?: () => void) =>
+          Promise.resolve().then(onfulfilled, onrejected),
+        cancel: jest.fn(),
+      };
+    }),
+  );
+});
+
+afterAll(() => {
+  jest.restoreAllMocks();
+});
+
 function renderAppInformation() {
   const state = initialStateIdentity().build();
-  return renderComponentViewScreen(
+  return renderScreenWithRoutes(
     AppInformation as unknown as React.ComponentType,
     { name: Routes.SETTINGS.COMPANY },
+    [{ name: Routes.WEBVIEW.MAIN }],
     { state },
   );
 }
@@ -22,49 +57,31 @@ describeForPlatforms('AppInformation (contact-us) component views', () => {
     jest.clearAllMocks();
   });
 
-  // Migrated from contact-us.spec.ts
-  // The original E2E test only navigated to the support screen and verified it opened.
-  // Here we verify the screen renders correctly and the "Contact us" link is present.
-  it('renders the About MetaMask screen with a Contact us link', async () => {
+  it('navigates to the webview when Contact us is pressed', async () => {
     const { findByTestId, findByText } = renderAppInformation();
 
-    expect(
-      await findByTestId(AboutMetaMaskSelectorsIDs.CONTAINER),
-    ).toBeOnTheScreen();
+    await findByTestId(AboutMetaMaskSelectorsIDs.CONTAINER);
+    fireEvent.press(await findByText(strings('app_information.contact_us')));
 
-    expect(
-      await findByText(strings('app_information.contact_us')),
-    ).toBeOnTheScreen();
+    await waitFor(async () => {
+      expect(
+        await findByTestId(getRouteProbeTestId(Routes.WEBVIEW.MAIN)),
+      ).toBeOnTheScreen();
+    });
   });
 
-  it('navigates to the webview when the Contact us link is pressed', async () => {
-    const { findByTestId, findByText, getByTestId } = renderAppInformation();
+  it('navigates to the webview when Support Center is pressed', async () => {
+    const { findByTestId, findByText } = renderAppInformation();
 
-    // Wait for the screen to be mounted
     await findByTestId(AboutMetaMaskSelectorsIDs.CONTAINER);
-
-    const contactUsLink = await findByText(
-      strings('app_information.contact_us'),
+    fireEvent.press(
+      await findByText(strings('app_information.support_center')),
     );
 
-    // Pressing the link triggers InteractionManager.runAfterInteractions then navigation.navigate.
-    // In CV tests we cannot assert real navigation to Webview (it's outside the rendered tree),
-    // but we can assert the element is pressable without throwing.
-    await waitFor(() => {
-      expect(contactUsLink).toBeOnTheScreen();
+    await waitFor(async () => {
+      expect(
+        await findByTestId(getRouteProbeTestId(Routes.WEBVIEW.MAIN)),
+      ).toBeOnTheScreen();
     });
-
-    fireEvent.press(contactUsLink);
-
-    // If navigation throws the test would fail; reaching here means the handler executed cleanly.
-    expect(contactUsLink).toBeOnTheScreen();
-  });
-
-  it('renders the Support Center link', async () => {
-    const { findByText } = renderAppInformation();
-
-    expect(
-      await findByText(strings('app_information.support_center')),
-    ).toBeOnTheScreen();
   });
 });
